django/requirements.txt
django
gunicorn
python-dotenv
motor
bs4
lxml
psycopg2
telethon
schedule
requests
pillow
pytelegrambotapi
celery
redis
django-solo
cryptography
django-hosts
django-cors-headers
xlwt
django-modeltranslation
django-redis
django-debug-toolbar
python-decouple
djangorestframework_simplejwt
djangorestframework
markdown
django-filter
